Initial Commit
Wanted to get this uploaded so people can check it out. 

Still a work in progress , but this is probably one of my most heavily commented tutorial trainers to date. 
I plan to complete my playthrough of FFVII as well as finish this trainer and comment on it a bunch more.

Stay tuned for updates.

What Works:
- GENERAL PANEL
- AWARD SCREEN PANEL
- ITEMS 
- BATTLE SCREEN PANEL (SLOTS 1 and 2)

What Doesn't Work:
- MATERIA
- BATTLE SCREEN PANEL (Slot 3)
- CHARACTERS PANEL (this is currently empty)